Moved shared web functionality to FunderMaps.AspNetCore
