+++ src/FunderMaps.AspNetCore/Extensions/ConfigurationExtensions.cs
