+++ src/FunderMaps.AspNetCore/Extensions/ExceptionMapperServiceCollectionExtensions.cs
@@ -1,10 +1,10 @@
-﻿using FunderMaps.AspNetCore.ErrorMessaging;
-using FunderMaps.Core.Exceptions;
+﻿using FunderMaps.Core.Exceptions;
 using FunderMaps.Core.Interfaces;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace FunderMaps.AspNetCore.Extensions
 {
+    // TODO This is a duplicate of FunderMaps.WebApi.Extensions.ExceptionMapperServiceCollectionExtensions.cs
     /// <summary>
     ///     Extension functionality to add exception mapping to the services.
     /// </summary>
@@ -15,7 +15,7 @@ namespace FunderMaps.AspNetCore.Extensions
         /// </summary>
         /// <param name="services"><see cref="IServiceCollection"/></param>
         /// <returns><see cref="IServiceCollection"/></returns>
-        public static IServiceCollection AddFunderMapsExceptionMapper(this IServiceCollection services) 
+        public static IServiceCollection AddFunderMapsExceptionMapper(this IServiceCollection services)
             => services.AddSingleton<IExceptionMapper<FunderMapsCoreException>, FunderMapsExceptionMapper>();
     }
 }
