+++ tests/FunderMaps.IntegrationTests/FunderMaps.IntegrationTests.csproj
@@ -24,6 +24,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\src\FunderMaps.AspNetCore\FunderMaps.AspNetCore.csproj" />
     <ProjectReference Include="..\..\src\FunderMaps.WebApi\FunderMaps.WebApi.csproj" />
     <ProjectReference Include="..\..\src\FunderMaps.Webservice\FunderMaps.Webservice.csproj" />
   </ItemGroup>
